Move schema_like to highlevel
