+++ tiledb/util.py
@@ -1,9 +1,7 @@
 from collections import deque
-from typing import Any, Iterable, Optional, Tuple, Union
 
 import numpy as np
 
-import tiledb
 import tiledb.cc as lt
 
 _dtype_to_tiledb = {
@@ -161,7 +159,7 @@ def array_type_ncells(dtype: np.dtype) -> lt.DataType:
     return tdb_type, ncells
 
 
-def dtype_range(dtype: np.dtype) -> Tuple[Any]:
+def dtype_range(dtype: np.dtype):
     """Return the range of a Numpy dtype"""
 
     if np.issubdtype(dtype, np.integer):
@@ -231,223 +229,3 @@ def numpy_dtype(tiledb_dtype: lt.DataType, cell_size: int = 1) -> np.dtype:
         return np.dtype([("", base_dtype)] * cell_val_num)
 
     raise TypeError("tiledb datatype not understood")
-
-
-def sparse_array_from_numpy(
-    uri: str, array: np.array, ctx: Optional["tiledb.Ctx"] = None, **kw
-):
-    """
-    Implementation of tiledb.from_numpy for dense arrays. See documentation
-    of tiledb.from_numpy.
-    """
-    if not ctx:
-        ctx = tiledb.default_ctx()
-
-    mode = kw.pop("mode", "ingest")
-    timestamp = kw.pop("timestamp", None)
-
-    if mode not in ("ingest", "schema_only", "append"):
-        raise tiledb.TileDBError(f"Invalid mode specified ('{mode}')")
-
-    if mode in ("ingest", "schema_only"):
-        try:
-            with tiledb.Array.load_typed(uri):
-                raise tiledb.TileDBError(f"Array URI '{uri}' already exists!")
-        except tiledb.TileDBError:
-            pass
-
-    if mode == "append":
-        kw["append_dim"] = kw.get("append_dim", 0)
-        if tiledb.ArraySchema.load(uri).sparse:
-            raise tiledb.TileDBError("Cannot append to sparse array")
-
-    if mode in ("ingest", "schema_only"):
-        schema = _schema_like_numpy(array, ctx=ctx, **kw)
-        tiledb.Array.create(uri, schema)
-
-    if mode in ("ingest", "append"):
-        kw["mode"] = mode
-        with tiledb.open(uri, mode="w", ctx=ctx, timestamp=timestamp) as arr:
-            # <TODO> probably need better typecheck here
-            if array.dtype == object:
-                arr[:] = array
-            else:
-                arr.write_direct(np.ascontiguousarray(array), **kw)
-
-    return tiledb.DenseArray(uri, mode="r", ctx=ctx)
-
-
-def schema_like(
-    *args,
-    shape: Optional[tuple] = None,
-    dtype: Optional[np.dtype] = None,
-    ctx: Optional["tiledb.Ctx"] = None,
-    **kw,
-) -> "tiledb.ArraySchema":
-    """
-    Return an ArraySchema corresponding to a NumPy-like object or
-    `shape` and `dtype` kwargs. Users are encouraged to pass 'tile'
-    and 'capacity' keyword arguments as appropriate for a given
-    application.
-
-    :param A: NumPy array-like object, or TileDB reference URI, optional
-    :param tuple shape: array shape, optional
-    :param dtype: array dtype, optional
-    :param Ctx ctx: TileDB Ctx
-    :param kwargs: additional keyword arguments to pass through, optional
-    :return: tiledb.ArraySchema
-    """
-    if not ctx:
-        ctx = tiledb.default_ctx()
-
-    def is_ndarray_like(arr):
-        return hasattr(arr, "shape") and hasattr(arr, "dtype") and hasattr(arr, "ndim")
-
-    # support override of default dimension dtype
-    dim_dtype = kw.pop("dim_dtype", np.uint64)
-    if len(args) == 1:
-        arr = args[0]
-        if is_ndarray_like(arr):
-            tiling = _regularize_tiling(kw.pop("tile", None), arr.ndim)
-            schema = _schema_like_numpy(arr, tile=tiling, dim_dtype=dim_dtype, ctx=ctx)
-        else:
-            raise ValueError("expected ndarray-like object")
-    elif shape and dtype:
-        if np.issubdtype(np.bytes_, dtype):
-            dtype = np.dtype("S")
-        elif np.issubdtype(dtype, np.unicode_):
-            dtype = np.dtype("U")
-
-        ndim = len(shape)
-        tiling = _regularize_tiling(kw.pop("tile", None), ndim)
-
-        dims = []
-        for d in range(ndim):
-            # support smaller tile extents by kw
-            # domain is based on full shape
-            tile_extent = tiling[d] if tiling else shape[d]
-            domain = (0, shape[d] - 1)
-            dims.append(
-                tiledb.Dim(domain=domain, tile=tile_extent, dtype=dim_dtype, ctx=ctx)
-            )
-
-        att = tiledb.Attr(dtype=dtype, ctx=ctx)
-        dom = tiledb.Domain(*dims, ctx=ctx)
-        schema = tiledb.ArraySchema(ctx=ctx, domain=dom, attrs=(att,), **kw)
-    elif kw is not None:
-        raise ValueError
-    else:
-        raise ValueError(
-            "Must provide either ndarray-like object or 'shape' "
-            "and 'dtype' keyword arguments"
-        )
-
-    return schema
-
-
-def _schema_like_numpy(array: np.array, ctx: Optional["tiledb.Ctx"] = None, **kw):
-    """
-    Internal helper function for schema_like to create array schema from
-    NumPy array-like object.
-    """
-    if not ctx:
-        ctx = tiledb.default_ctx()
-    # create an ArraySchema from the numpy array object
-    tiling = _regularize_tiling(kw.pop("tile", None), array.ndim)
-
-    attr_name = kw.pop("attr_name", "")
-    dim_dtype = kw.pop("dim_dtype", np.dtype("uint64"))
-    full_domain = kw.pop("full_domain", False)
-    dims = []
-
-    for d in range(array.ndim):
-        # support smaller tile extents by kw
-        # domain is based on full shape
-        tile_extent = tiling[d] if tiling else array.shape[d]
-        if full_domain:
-            if dim_dtype not in (np.bytes_, np.str_):
-                # Use the full type domain, deferring to the constructor
-                dtype_min, dtype_max = dtype_range(dim_dtype)
-                dim_max = dtype_max
-                if dim_dtype.kind == "M":
-                    date_unit = np.datetime_data(dim_dtype)[0]
-                    dim_min = np.datetime64(dtype_min, date_unit)
-                    tile_max = np.iinfo(np.uint64).max - tile_extent
-                    if np.uint64(dtype_max - dtype_min) > tile_max:
-                        dim_max = np.datetime64(dtype_max - tile_extent, date_unit)
-                else:
-                    dim_min = dtype_min
-
-                if np.issubdtype(dim_dtype, np.integer):
-                    tile_max = np.iinfo(np.uint64).max - tile_extent
-                    if np.uint64(dtype_max - dtype_min) > tile_max:
-                        dim_max = dtype_max - tile_extent
-                domain = (dim_min, dim_max)
-            else:
-                domain = (None, None)
-
-            if np.issubdtype(dim_dtype, np.integer) or dim_dtype.kind == "M":
-                # we can't make a tile larger than the dimension range or lower than 1
-                tile_extent = max(1, min(tile_extent, np.uint64(dim_max - dim_min)))
-            elif np.issubdim_dtype(dim_dtype, np.floating):
-                # this difference can be inf
-                with np.errstate(over="ignore"):
-                    dim_range = dim_max - dim_min
-                if dim_range < tile_extent:
-                    tile_extent = np.ceil(dim_range)
-        else:
-            domain = (0, array.shape[d] - 1)
-
-        dims.append(
-            tiledb.Dim(domain=domain, tile=tile_extent, dtype=dim_dtype, ctx=ctx)
-        )
-
-    var = False
-    if array.dtype == object:
-        # for object arrays, we use the dtype of the first element
-        # consistency check should be done later, if needed
-        el0 = array.flat[0]
-        if type(el0) is bytes:
-            el_dtype = np.dtype("S")
-            var = True
-        elif type(el0) is str:
-            el_dtype = np.dtype("U")
-            var = True
-        elif type(el0) == np.ndarray:
-            if len(el0.shape) != 1:
-                raise TypeError(
-                    "Unsupported sub-array type for Attribute: {} "
-                    "(only string arrays and 1D homogeneous NumPy arrays are supported)".format(
-                        type(el0)
-                    )
-                )
-            el_dtype = el0.dtype
-        else:
-            raise TypeError(
-                "Unsupported sub-array type for Attribute: {} "
-                "(only strings and homogeneous-typed NumPy arrays are supported)".format(
-                    type(el0)
-                )
-            )
-    else:
-        el_dtype = array.dtype
-
-    att = tiledb.Attr(dtype=el_dtype, name=attr_name, var=var, ctx=ctx)
-    dom = tiledb.Domain(*dims, ctx=ctx)
-    return tiledb.ArraySchema(ctx=ctx, domain=dom, attrs=(att,), **kw)
-
-
-def _regularize_tiling(tile: Union[Iterable, np.isscalar], ndim: int) -> Tuple[Any]:
-    """
-    Internal helper function for schema_like and schema_like_numpy to regularize tiling.
-    """
-    if not tile:
-        return None
-
-    if np.isscalar(tile):
-        return tuple(int(tile) for _ in range(ndim))
-
-    if isinstance(tile, str) or len(tile) != ndim:
-        raise ValueError("'tile' must be iterable and match array dimensionality")
-
-    return tuple(tile)
